This will help people get started
First, I am no longer using a debug database so that's commented out.

But the big change here is from setting the userTableName from "UserProfile" to "Users".

This change allows the people who created the database from the .edmx  file (like I instructed) to not run into issues because that database is called "Users" not "UserProfiles"

My humble apologies.